% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_Hd}
\alias{add_errors_Hd}
\title{Add Hp-errors to table}
\usage{
add_errors_Hd(tab, w)
}
\arguments{
\item{tab}{table to add errors to}

\item{w}{error probability}
}
\description{
Add Hp-errors to table
}
\examples{
Z_D <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
Z_S <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
tab <- table(to012(Z_D), to012(Z_S))
tab
new_tab <- add_errors_Hd(tab, w = 0.15)
new_tab
estimate_w(new_tab) # wrong!
Z <- sample_profiles_without_error(n = 10000, p = c(0.25, 0.25, 0.5))
tab <- table(to012(Z), to012(Z))
tab
new_tab <- add_errors_Hd(tab, w = 0.25)
new_tab
estimate_w(new_tab) # ok

}
