% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_Hd_wDwS}
\alias{add_errors_Hd_wDwS}
\title{Add Hd-errors to table for sample-dependent error probabilities, $w_D$ and $w_S$}
\usage{
add_errors_Hd_wDwS(tab, wD, wS)
}
\arguments{
\item{tab}{table to add errors to}

\item{wD}{error probability for donor sample}

\item{wS}{error probability for PoI sample}
}
\description{
Add Hd-errors to table for sample-dependent error probabilities, $w_D$ and $w_S$
}
\examples{
Z_D <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
Z_S <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
tab <- table(to012(Z_D), to012(Z_S))
tab
add_errors_Hd_wDwS(tab, wD = 0.1, wS = 1e-6)
add_errors_Hd_w(tab, w = 0.05)

}
