% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_Hp_w}
\alias{add_errors_Hp_w}
\title{Add Hp-errors to table for one error probability, $w$}
\usage{
add_errors_Hp_w(tab, w)
}
\arguments{
\item{tab}{table to add errors to}

\item{w}{error probability}
}
\description{
Add Hp-errors to table for one error probability, $w$
}
\examples{
Z <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
tab <- table(to012(Z))
tab
new_tab <- add_errors_Hp_w(tab, w = 0.15)
new_tab
estimate_w(new_tab)

}
