% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_Hp_wTwR}
\alias{add_errors_Hp_wTwR}
\title{Add Hp-errors to table for sample-dependent error probabilities, $w_t$ and $w_r$}
\usage{
add_errors_Hp_wTwR(tab, wT, wR)
}
\arguments{
\item{tab}{table to add errors to}

\item{wT}{error probability for donor sample}

\item{wR}{error probability for PoI sample}
}
\description{
Add Hp-errors to table for sample-dependent error probabilities, $w_t$ and $w_r$
}
\examples{
Z <- sample_profiles_without_error(n = 1000, p = c(0.25, 0.25, 0.5))
tab <- table(to012(Z))
tab
add_errors_Hp_wTwR(tab, wT = 0.1, wR = 1e-6)
add_errors_Hp_w(tab, w = 0.05)

}
