% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_to_012genotypes}
\alias{add_errors_to_012genotypes}
\title{Add errors to 0/1/2 genotypes}
\usage{
add_errors_to_012genotypes(z, w)
}
\arguments{
\item{w}{error probability}

\item{Z}{genotypes in 0/1/2 format, e.g. created by \code{\link[=sample_profiles_without_error]{sample_profiles_without_error()}} followed by \code{\link[=to012]{to012()}}}
}
\value{
list, element for each locus is a matrix with n rows and two columns
}
\description{
Add errors to 0/1/2 genotypes
}
\examples{
Z <- sample_profiles_without_error(n = 100, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
X <- add_errors_to_genotypes(Z, 0.5)
table(to012(X))
X <- add_errors_to_012genotypes(to012(Z), 0.5)
table(X)
estimate_w(tab)

Z012 <- to012(Z)
X1 <- add_errors_to_012genotypes(Z012, 0.1)
X2 <- add_errors_to_012genotypes(Z012, 0.1)
tab <- table(X1, X2)
tab
estimate_w(tab)

}
