% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{allele_probs_to_geno_probs}
\alias{allele_probs_to_geno_probs}
\title{Allele probabilities to genotype probabilities assuming Hardy-Weinberg equilibrium}
\usage{
allele_probs_to_geno_probs(x)
}
\description{
Allele probabilities to genotype probabilities assuming Hardy-Weinberg equilibrium
}
\examples{
allele_probs_to_geno_probs(c(0.1, 0.2))

}
