% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-integrate.R
\name{alt_calc_WoE_wTwR_integrate_wT_mc}
\alias{alt_calc_WoE_wTwR_integrate_wT_mc}
\title{Alternative calculation of WoE for a profile for sample-specific error probabilities integrated over the donor prior distribution using Monte Carlo integration}
\usage{
alt_calc_WoE_wTwR_integrate_wT_mc(
  xT,
  xR,
  shape1T,
  shape2T,
  wR,
  p,
  n_samples = 1000
)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{shape1T}{under $H_1$ \code{wT} has beta prior on (0, 0.5) with parameters \code{shape1T_H1} and \code{shape2T_H1}}

\item{shape2T}{see \code{shape1T_H1}}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}

\item{n_samples}{number of random samples from each prior distribution}
}
\description{
This does not calculate the LR as expectation of the product over markers as [], but instead calculate the
expectation of the log of the product over markers.
}
\details{
Note that the same prior distribution is assumed under $H_p$ and $H_a$.
}
\examples{
calc_LRs_wTwR(xT = c(0, 0), xR = c(0, 1), wT = 1e-2, wR = 1e-5, p = c(0.25, 0.25, 0.5)) |> log10() |> sum()

shpT <- get_beta_parameters(mu = 1e-2, sigmasq = 1e-7, a = 0, b = 0.5)
# curve(dbeta05(x, shpT[1], shpT[2]), from = 0, to = 0.1, n = 1001)
z1 <- calc_WoE_wTwR_integrate_wT_mc(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1T_Hp = shpT[1], 
  shape2T_Hp = shpT[2],
  shape1T_Ha = shpT[1], 
  shape2T_Ha = shpT[2],
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 1000)
z1

z2 <- alt_calc_WoE_wTwR_integrate_wT_mc(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1T = shpT[1], shape2T = shpT[2], 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5))
z2
  
}
