% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{rbeta05}
\alias{rbeta05}
\alias{pbeta05}
\alias{qbeta05}
\alias{dbeta05}
\title{Beta distribution on (0, 0.5)}
\usage{
rbeta05(n, shape1, shape2)

pbeta05(x, shape1, shape2)

qbeta05(x, shape1, shape2)

dbeta05(x, shape1, shape2, log = FALSE)
}
\arguments{
\item{n}{number of random samples to draw}

\item{shape1}{first shape parameter}

\item{shape2}{second shape parameter}

\item{x}{number between 0 and 0.5}

\item{log}{return result on log scale}
}
\description{
Beta distribution on (0, 0.5)
}
\examples{
dbeta05(0.2, 1, 5)
dbeta05(0.2, 1, 5, log = TRUE)
dbeta05(0.2, 1, 5) |> log()

pbeta05(0.5, 1, 5)

qbeta05(1, 1, 5)

#rbeta05(100, 1, 5) |> hist(probability = TRUE)
#curve(dbeta05(x, 1, 5), from = 0, to = 0.5, add = TRUE)

}
