% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-integrate.R
\name{calc_LRs_wDwS_integrate_wDwS_mc}
\alias{calc_LRs_wDwS_integrate_wDwS_mc}
\title{Calculate LR for a profile for sample-specific error probabilities integrated over the prior distributions using Monte Carlo integration}
\usage{
calc_LRs_wDwS_integrate_wDwS_mc(
  xD,
  xS,
  shape1D,
  shape2D,
  shape1S,
  shape2S,
  p,
  n_samples = 1000
)
}
\arguments{
\item{xD}{profile from case (of 0, 1, 2)}

\item{xS}{profile from suspect (of 0, 1, 2)}

\item{shape1D}{\code{wD} has beta prior (with support on 0-0.5) with parameters \code{shape1D} and \code{shape2D}}

\item{shape2D}{see \code{shape1D}}

\item{shape1S}{\code{wS} has beta prior (with support on 0-0.5) with parameters \code{shape1S} and \code{shape2S}}

\item{shape2S}{see \code{shape1S}}

\item{p}{list of genotype probabilities (same length as \code{xD}/\code{xS}, or vector of length 3 for reuse)}

\item{n_samples}{number of random samples from each prior distribution}
}
\description{
Calculate LR for a profile for sample-specific error probabilities integrated over the prior distributions using Monte Carlo integration
}
\examples{
calc_LRs_wDwS(xD = c(0, 0), xS = c(0, 1), wD = 1e-2, wS = 1e-5, p = c(0.25, 0.25, 0.5))

shpD <- get_beta_parameters(mu = 1e-2, sigmasq = 2*1e-3, a = 0, b = 0.5)
# curve(dbeta05(x, shpD[1L], shpD[2L]), from = 0, to = 0.5)
shpS <- get_beta_parameters(mu = 1e-5, sigmasq = 2e-6, a = 0, b = 0.5)
# curve(dbeta05(x, shpS[1L], shpS[2L]), from = 0, to = 0.5)

calc_LRs_wDwS_integrate_wDwS_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  shape1S = shpS[1], shape2S = shpS[2], 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
  
calc_LRs_wDwS_integrate_wDwS_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = 1, shape2D = 1, 
  shape1S = 1, shape2S = 1, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, 1, 1), from = 0, to = 0.5)

}
