% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{calc_LRs_wDwS_uncertainty}
\alias{calc_LRs_wDwS_uncertainty}
\title{Calculate LR for a profile for sample-specific error probabilities}
\usage{
calc_LRs_wDwS_uncertainty(
  xD,
  xS,
  shape1D,
  shape2D,
  shape1S,
  shape2S,
  p,
  method = c("integrate", "simulate")
)
}
\arguments{
\item{xD}{profile from case (of 0, 1, 2)}

\item{xS}{profile from suspect (of 0, 1, 2)}

\item{shape1D}{\code{wD} has beta prior (with support on 0-0.5) with parameters \code{shape1D} and \code{shape2D}}

\item{shape2D}{see \code{shape1D}}

\item{shape1S}{\code{wS} has beta prior (with support on 0-0.5) with parameters \code{shape1S} and \code{shape2S}}

\item{shape2S}{see \code{shape1S}}

\item{p}{list of genotype probabilities (same length as \code{xD}/\code{xS}, or vector of length 3 for reuse)}

\item{method}{numerically \code{integrate} or \code{simulate} (using 100 values for \code{wD} and 100 values for \code{wS})}
}
\description{
Calculate LR for a profile for sample-specific error probabilities
}
\examples{
shpD <- get_beta_parameters(mu = 1e-2, sigmasq = 1e-3, a = 0, b = 0.5)
shpS <- get_beta_parameters(mu = 1e-5, sigmasq = 1e-8, a = 0, b = 0.5)
integrate(function(x) dbeta05(x, shape1 = shpD[1L], shape2 = shpD[2L]), lower = 0, upper = 0.5, rel.tol = 1e-6)
integrate(function(x) dbeta05(x, shape1 = shpS[1L], shape2 = shpS[2L]), lower = 0, upper = 0.5, rel.tol = 1e-6)

calc_LRs_wDwS(c(0, 0), c(0, 1), wD = 1e-2, wS = 1e-6, p = c(0.25, 0.25, 0.5))

calc_LRs_wDwS_uncertainty(c(0, 0), c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  shape1S = shpS[1], shape2S = shpS[2], 
  p = c(0.25, 0.25, 0.5), 
  method = "integrate")
calc_LRs_wDwS_uncertainty(c(0, 0), c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  shape1S = shpS[1], shape2S = shpS[2], 
  p = c(0.25, 0.25, 0.5), 
  method = "simulate")

}
