% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{calc_LRs_wTwR}
\alias{calc_LRs_wTwR}
\title{Calculate LR for a profile for sample-specific error probabilities}
\usage{
calc_LRs_wTwR(xT, xR, wT, wR, p)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{wT}{error probability for donor sample}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}
}
\description{
Calculate LR for a profile for sample-specific error probabilities
}
\examples{
calc_LRs_wTwR(c(0, 0), c(0, 0), wT = 0, wR = 0, p = c(0.25, 0.25, 0.5))
calc_LRs_w(c(0, 0), c(0, 0), w = 0, p = c(0.25, 0.25, 0.5))

calc_LRs_wTwR(c(0, 0), c(0, 0), wT = 0, wR = 0, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
calc_LRs_w(c(0, 0), c(0, 0), w = 0, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
  
calc_LRs_wTwR(c(0, 0), c(0, 1), wT = 0, wR = 0, p = c(0.25, 0.25, 0.5))

calc_LRs_wTwR(c(0, 0), c(0, 1), wT = 1e-5, wR = 1e-5, p = c(0.25, 0.25, 0.5))
calc_LRs_w(c(0, 0), c(0, 1), w = 1e-5, p = c(0.25, 0.25, 0.5))

calc_LRs_wTwR(c(0, 0), c(0, 1), wT = 1e-3, wR = 1e-3, p = c(0.25, 0.25, 0.5))
calc_LRs_w(c(0, 0), c(0, 1), w = 1e-3, p = c(0.25, 0.25, 0.5))

calc_LRs_wTwR(c(0, 0), c(0, 1), wT = 1e-2, wR = 1e-6, p = c(0.25, 0.25, 0.5))
calc_LRs_w(c(0, 0), c(0, 1), w = 5e-3, p = c(0.25, 0.25, 0.5))

}
