% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-integrate.R
\name{calc_LRs_wTwR_integrate_wT}
\alias{calc_LRs_wTwR_integrate_wT}
\title{Calculate LR for a profile for sample-specific error probabilities integrated over the donor prior distribution using exact integration}
\usage{
calc_LRs_wTwR_integrate_wT(
  xT,
  xR,
  shape1T_H1,
  shape2T_H1,
  shape1T_H2,
  shape2T_H2,
  wR,
  p,
  use_mpfr = TRUE,
  stop_on_infinite = TRUE,
  mpfr_precision = 256
)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{shape1T_H1}{Under $H_1$ (in $LR$'s numerator), \code{wT} has beta prior on (0, 0.5) with parameters \code{shape1T_H1} and \code{shape2T_H1}}

\item{shape2T_H1}{see \code{shape1T_Hp}}

\item{shape1T_H2}{Under $H_2$ (in $LR$'s denominator), \code{wT} has beta prior (on 0-0.5) with parameters \code{shape1T_H2} and \code{shape2T_H2}}

\item{shape2T_H2}{see \code{shape1T_H2}}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}

\item{use_mpfr}{use higher precision numbers via the \code{Rmpfr} package}

\item{stop_on_infinite}{stop if infinite numbers are encountered (if so, try \code{use_mpfr = TRUE})}

\item{mpfr_precision}{number of bits to use}
}
\description{
NOTE: Unless you have a reason, please use \code{\link[=calc_LRs_wTwR_integrate_wT]{calc_LRs_wTwR_integrate_wT()}} instead.
}
\examples{
calc_LRs_wTwR(xT = c(0, 0), xR = c(0, 1), wT = 1e-2, wR = 1e-5, p = c(0.25, 0.25, 0.5))

shpD <- get_beta_parameters(mu = 1e-2, sigmasq = 2*1e-3, a = 0, b = 0.5)
calc_LRs_wTwR_integrate_wT_mc(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
calc_LRs_wTwR_integrate_wT_num(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5))
# curve(dbeta05(x, shpD[1], shpD[2]), from = 0, to = 0.5)

calc_LRs_wTwR_integrate_wT_mc(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1D = 1, shape2D = 1, 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
calc_LRs_wTwR_integrate_wT_num(
  xT = c(0, 0),
  xR = c(0, 1),
  shape1D = 1, shape2D = 1,
  wR = 1e-5,
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, 1, 1), from = 0, to = 0.5)
  
}
