% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-integrate.R
\name{calc_WoE_wTwR_integrate_wT_mc_markerwise}
\alias{calc_WoE_wTwR_integrate_wT_mc_markerwise}
\title{Calculate WoE for a profile for sample-specific error probabilities integrated over the donor prior distribution using Monte Carlo integration}
\usage{
calc_WoE_wTwR_integrate_wT_mc_markerwise(
  xT,
  xR,
  shape1T_H1,
  shape2T_H1,
  shape1T_H2,
  shape2T_H2,
  wR,
  p,
  n_samples = 1000
)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{shape1T_H1}{Under $H_1$ (in $LR$'s numerator), \code{wT} has beta prior on (0, 0.5) with parameters \code{shape1T_H1} and \code{shape2T_H1}}

\item{shape2T_H1}{see \code{shape1T_Hp}}

\item{shape1T_H2}{Under $H_2$ (in $LR$'s denominator), \code{wT} has beta prior (on 0-0.5) with parameters \code{shape1T_H2} and \code{shape2T_H2}}

\item{shape2T_H2}{see \code{shape1T_H2}}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}

\item{n_samples}{number of random samples from each prior distribution}
}
\description{
Note that the WoE for each marker is calculated by integrating over the prior of $w_t$
separately under $H_1$ and $H_2$.
An alternative is implemented in \code{\link[=calc_WoE_wTwR_integrate_wT_mc]{calc_WoE_wTwR_integrate_wT_mc()}} as
the expected value of log10 of $LR$ under a prior of $w_t$.
}
\details{
Results from this can be compared to those from \code{\link[=calc_WoE_wTwR_integrate_wT_exact_markerwise]{calc_WoE_wTwR_integrate_wT_exact_markerwise()}}.
}
\examples{
calc_LRs_wTwR(xT = c(0, 0), xR = c(0, 1), wT = 1e-2, wR = 1e-5, p = c(0.25, 0.25, 0.5)) |> log10()

shpT <- get_beta_parameters(mu = 1e-2, sigmasq = 1e-5, a = 0, b = 0.5)
# curve(dbeta05(x, shpT[1], shpT[2]), from = 0, to = 0.1, n = 1001)
calc_WoE_wTwR_integrate_wT_mc_markerwise(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1T_H1 = shpT[1], shape2T_H1 = shpT[2],
  shape1T_H2 = shpT[1], shape2T_H2 = shpT[2],
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 1000)
  
calc_WoE_wTwR_integrate_wT_exact_markerwise(
  xT = c(0, 0), 
  xR = c(0, 1), 
  shape1T_H1 = shpT[1], shape2T_H1 = shpT[2],
  shape1T_H2 = shpT[1], shape2T_H2 = shpT[2],
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5))

}
