% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR-integrate.R
\name{calc_WoE_wTwR_mleH2_wT_num}
\alias{calc_WoE_wTwR_mleH2_wT_num}
\title{Calculate WoE for sample-specific error probabilities using maximum likelihood of \code{wT} under $H_2$ using numerical optimisation}
\usage{
calc_WoE_wTwR_mleH2_wT_num(
  xT,
  xR,
  wR,
  p,
  opttype = c("optimise_strict", "optimise_nocheck", "DEoptim")
)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}

\item{opttype}{Optimiser type \code{"optimise_strict"}, \code{"optimise_nocheck"}, or \code{"DEoptim"}}
}
\description{
Calculate WoE for sample-specific error probabilities using maximum likelihood of \code{wT} under $H_2$ using numerical optimisation
}
\examples{
calc_LRs_wTwR(
  xT = c(0, 0), 
  xR = c(0, 1), 
  wT = 1e-2, 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5)) |> log10() |> sum()
  
calc_WoE_wTwR_mleH2_wT_num(
  xT = c(0, 0), 
  xR = c(0, 1), 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5))
  
calc_WoE_wTwR_profilemax_wT_num(
  xT = c(0, 0), 
  xR = c(0, 1), 
  wR = 1e-5, 
  p = c(0.25, 0.25, 0.5))
  
}
