% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_formulas_Ha_w}
\alias{d_formulas_Ha_w}
\title{Enumeration of possibilities under Hd for one error probability, $w$}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 20 columns.
}
\usage{
d_formulas_Ha_w
}
\description{
Enumeration of possibilities under Hd for one error probability, $w$
}
\keyword{datasets}
