% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_formulas_Hd_wTwR}
\alias{d_formulas_Hd_wTwR}
\title{Enumeration of possibilities under Hd for sample-dependent error probabilities, $w_t$ and $w_r$}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 24 columns.
}
\usage{
d_formulas_Hd_wTwR
}
\description{
Enumeration of possibilities under Hd for sample-dependent error probabilities, $w_t$ and $w_r$
}
\keyword{datasets}
