% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_formulas_Hp}
\alias{d_formulas_Hp}
\title{Enumeration of possibilities under Hp}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 16 columns.
}
\usage{
d_formulas_Hp
}
\description{
Enumeration of possibilities under Hp
}
\keyword{datasets}
