% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_formulas_Hp_wDwS}
\alias{d_formulas_Hp_wDwS}
\title{Enumeration of possibilities under Hp for sample-dependent error probabilities, $w_D$ and $w_S$}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 20 columns.
}
\usage{
d_formulas_Hp_wDwS
}
\description{
Enumeration of possibilities under Hp for sample-dependent error probabilities, $w_D$ and $w_S$
}
\keyword{datasets}
