% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_formulas_single}
\alias{d_formulas_single}
\title{Enumeration of possibilities for single observation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 11 columns.
}
\usage{
d_formulas_single
}
\description{
Enumeration of possibilities for single observation
}
\keyword{datasets}
