% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_LR_w}
\alias{d_LR_w}
\title{Formulas for LR grouped by $X^D$ and $X^S$ for one error probability, $w$}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
d_LR_w
}
\description{
Formulas for LR grouped by $X^D$ and $X^S$ for one error probability, $w$
}
\keyword{datasets}
