% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_prob_Ha_wTwR}
\alias{d_prob_Ha_wTwR}
\title{Formulas under Hd grouped by $X^D$ and $X^S$ for sample-dependent error probabilities, $w_t$ and $w_r$}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
d_prob_Ha_wTwR
}
\description{
Formulas under Hd grouped by $X^D$ and $X^S$ for sample-dependent error probabilities, $w_t$ and $w_r$
}
\keyword{datasets}
