% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_prob_Hd}
\alias{d_prob_Hd}
\title{Formulas under Hd grouped by XD and XS}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
d_prob_Hd
}
\description{
Formulas under Hd grouped by XD and XS
}
\keyword{datasets}
