% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_prob_Hp_wDwS}
\alias{d_prob_Hp_wDwS}
\title{Formulas under Hp grouped by XD and XS for sample-dependent error probabilities, $w_D$ and $w_S$}
\format{
An object of class \code{rowwise_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 5 columns.
}
\usage{
d_prob_Hp_wDwS
}
\description{
Formulas under Hp grouped by XD and XS for sample-dependent error probabilities, $w_D$ and $w_S$
}
\keyword{datasets}
