% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_probtable_Ha_w}
\alias{d_probtable_Ha_w}
\title{Probabilties for efficient sampling under Hd for one error probability, $w$}
\format{
An object of class \code{data.frame} with 256 rows and 13 columns.
}
\usage{
d_probtable_Ha_w
}
\description{
Probabilties for efficient sampling under Hd for one error probability, $w$
}
\keyword{datasets}
