% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_probtable_Hd}
\alias{d_probtable_Hd}
\title{Probabilties for efficient sampling under Hd}
\format{
An object of class \code{data.frame} with 256 rows and 13 columns.
}
\usage{
d_probtable_Hd
}
\description{
Probabilties for efficient sampling under Hd
}
\keyword{datasets}
