% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_probtable_Hp}
\alias{d_probtable_Hp}
\title{Probabilties for efficient sampling under Hp}
\format{
An object of class \code{data.frame} with 64 rows and 9 columns.
}
\usage{
d_probtable_Hp
}
\description{
Probabilties for efficient sampling under Hp
}
\keyword{datasets}
