% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_probtable_Hp_wTwR}
\alias{d_probtable_Hp_wTwR}
\title{Probabilties for efficient sampling under Hp for sample-dependent error probabilities, $w_t$ and $w_r$}
\format{
An object of class \code{data.frame} with 64 rows and 9 columns.
}
\usage{
d_probtable_Hp_wTwR
}
\description{
Probabilties for efficient sampling under Hp for sample-dependent error probabilities, $w_t$ and $w_r$
}
\keyword{datasets}
