% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{dbeta05}
\alias{dbeta05}
\title{dbeta05(0.2, 1, 5)
dbeta05(0.2, 1, 5, log = TRUE)
dbeta05(0.2, 1, 5) |>}
\usage{
dbeta05(x, shape1, shape2, log = FALSE)
}
\description{
dbeta05(0.2, 1, 5)
dbeta05(0.2, 1, 5, log = TRUE)
dbeta05(0.2, 1, 5) |>
}
