% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_w}
\alias{estimate_w}
\title{Estimate error probability, w}
\usage{
estimate_w(x, use_mpfr = FALSE, ...)
}
\arguments{
\item{x}{either a 3x3 contingency table or
3-vector with counts (\code{n1}, \code{n2}, \code{n3})
where \code{n1 = sum(diag(x))}, \code{n2 = x[1L, 3L] + x[3L, 1L]},
\code{n3 = sum(x) - n1 - n2}}

\item{use_mpfr}{whether to use \code{\link[Rmpfr:optimizeR]{Rmpfr::optimizeR()}}] for arbitrary precision}

\item{\dots}{passed on to \code{optimise()} or to \code{Rmpfr::optimizeR} if \code{use_mpfr = TRUE}}
}
\description{
Note that \code{tol = .Machine$double.eps} unless overwritten.
}
\examples{
tab1 <- matrix(c(1000, 10, 2, 12, 100, 8, 1, 7, 200), nrow = 3)
tab1
estimate_w(tab1)
n1 <- sum(diag(tab1))
n2 <- tab1[1L, 3L] + tab1[3L, 1L]
n3 <- sum(tab1) - n1 - n2
estimate_w(c(n1, n2, n3))

tab2 <- structure(c(40000L, 30L, 5L, 15L, 400L, 2L, 5L, 5L, 350L),
        dim = c(3L, 3L), class = "table")
tab2
estimate_w(tab2)
estimate_w(tab2, use_mpfr = TRUE)

}
