% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_w_bayesian}
\alias{estimate_w_bayesian}
\title{Estimate error probability, w, using a Bayesian approach}
\usage{
estimate_w_bayesian(
  x,
  prior_a = 1,
  prior_b = 1,
  chains = 4,
  chains_function = lapply,
  warmup = 1000,
  iterations = 10000,
  ...
)
}
\arguments{
\item{x}{either a 3x3 contingency table or
3-vector with counts (\code{n1}, \code{n2}, \code{n3})
where \code{n1 = sum(diag(x))}, \code{n2 = x[1L, 3L] + x[3L, 1L]},
\code{n3 = sum(x) - n1 - n2}}

\item{prior_a}{First shape parameter for prior beta distribution}

\item{prior_b}{Second shape parameter for prior beta distribution}

\item{iterations}{Number of iterations}

\item{\dots}{Not used}
}
\description{
A Beta prior is assumed. Note that $w$ is assumed to be $< 0.5$.
}
\details{
NOTE: We recommend using e.g. the \code{cmdstanr}, see
the vignette. This is just an implementation that demonstrate the
Bayesian approach.
}
\examples{
tab1 <- matrix(c(1000, 10, 2, 12, 100, 8, 1, 7, 200), nrow = 3)
tab1
estimate_w(tab1)
y <- estimate_w_bayesian(tab1)
q <- posterior_samples(y)
mean(q); #plot(q, type = "l"); hist(q)
sapply(y, \(x) x$acceptance_ratio)

estimate_w_bayesian(c(750, 0, 0), chains = 1, warmup = 0, iterations = 500) |> 
   lapply(\(x) x$samples) |> unlist() |> plot(type = "l")

estimate_w(c(750, 0, 0))
y <- estimate_w_bayesian(c(750, 0, 0)); q <- lapply(y, \(x) x$samples) |> unlist()
mean(q); #plot(q, type = "l"); hist(q)
sapply(y, \(x) x$acceptance_ratio)

}
