% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-profile.R
\name{extract_profile}
\alias{extract_profile}
\title{Extract profile from ranked markers}
\usage{
extract_profile(x, markers = marker_candidates_v1)
}
\arguments{
\item{x}{Data (potentially already filtered), including columns \code{chr} and \code{pos}}

\item{markers}{Marker ranking}
}
\value{
Original data with extra columns \code{segment}, \code{segment_no} and \code{rank}
}
\description{
Uses the most recent marker ranking (see function argument).
}
\details{
NOTE: This does not check if reference and alternative alleles are the same.
See examples below for one way to inspect this.
}
\examples{
vcf <- vcf_example_data_v1 |> rename(chr = CHROM, pos = POS)
prof <- extract_profile(vcf)
head(prof)
nrow(prof)
table(prof$rank)

# Also using allele frequencies from `marker_candidates_v1`
prof_extra_info <- prof |> inner_join(marker_candidates_v1 |> 
  select(chr, pos, ref, alt, AF_afr, AF_eas, AF_nfe), by = c("chr", "pos"))
head(prof_extra_info)

subset(prof_extra_info, ref != REF)
subset(prof_extra_info, alt != ALT)
subset(prof_extra_info, alt != ALT & ALT != "?")

X <- to012(prof_extra_info$GT)
X
genoprob <- allele_probs_to_geno_probs(prof_extra_info$AF_nfe)

LRs <- calc_LRs(xs = X, xd = X, w = 0.01, p = genoprob)
WoE <- sum(log10(LRs))
WoE

ws <- 10^(-(6:1))
LRs_ws <- sapply(ws, \(w) sum(log10(calc_LRs(xs = X, xd = X, w = w, p = genoprob))))
cbind(ws, LRs_ws)

# Assume two errors
head(X)
Xd <- X
Xd[1] <- 1
Xd[2] <- 1
LRs_ws_Xd <- sapply(ws, \(w) sum(log10(calc_LRs(xs = X, xd = Xd, w = w, p = genoprob))))
cbind(ws, LRs_ws_Xd)

# Assume two more errors
head(X)
Xd[3] <- 0
Xd[4] <- 0
LRs_ws_Xd4 <- sapply(ws, \(w) sum(log10(calc_LRs(xs = X, xd = Xd, w = w, p = genoprob))))
cbind(ws, LRs_ws_Xd4)

}
