% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-profile.R
\name{extract_profile_v1}
\alias{extract_profile_v1}
\title{Extract profile}
\usage{
extract_profile_v1(x)
}
\arguments{
\item{x}{Data (potentially already filtered), including columns \code{chr} and \code{pos}}
}
\description{
Uses \code{marker_candidates_v1} ranking.
}
\examples{
vcf <- vcf_example_data_v1 |> rename(chr = CHROM, pos = POS)
prof <- extract_profile(vcf)
head(prof)
nrow(prof)
table(prof$rank)

# Also using allele frequencies from `marker_candidates_v1`
prof_extra_info <- prof |> inner_join(marker_candidates_v1 |> 
  select(chr, pos, ref, alt, AF_afr, AF_eas, AF_nfe), by = c("chr", "pos"))
head(prof_extra_info)

subset(prof_extra_info, ref != REF)
subset(prof_extra_info, alt != ALT)
subset(prof_extra_info, alt != ALT & ALT != "?")

X <- to012(prof_extra_info$GT)
X
genoprob <- allele_probs_to_geno_probs(prof_extra_info$AF_nfe)

LRs <- calc_LRs(xD = X, xS = X, w = 0.01, p = genoprob)
WoE <- sum(log10(LRs))
WoE

ws <- 10^(-(6:1))
LRs_ws <- sapply(ws, \(w) sum(log10(calc_LRs(xD = X, xS = X, w = w, p = genoprob))))
cbind(ws, LRs_ws)

# Assume two errors
head(X)
XD <- X
XD[1] <- 1
XD[2] <- 1
LRs_ws_Xd <- sapply(ws, \(w) sum(log10(calc_LRs(xD = XD, xS = X, w = w, p = genoprob))))
cbind(ws, LRs_ws_Xd)

# Assume two more errors
head(X)
XD[3] <- 0
XD[4] <- 0
LRs_ws_Xd4 <- sapply(ws, \(w) sum(log10(calc_LRs(xD = XD, xS = X, w = w, p = genoprob))))
cbind(ws, LRs_ws_Xd4)

}
