% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{get_beta_mu_sigma}
\alias{get_beta_mu_sigma}
\title{Parameterisation conversion for beta distribution}
\usage{
get_beta_mu_sigma(shape1, shape2, a = 0, b = 0.5)
}
\arguments{
\item{shape1}{first shape parameter}

\item{shape2}{second shape parameter}

\item{a}{lower support bound}

\item{b}{upper support bound}
}
\description{
Get beta distribution's mean and variance.
}
