% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_Ha_wTwR}
\alias{graph_Ha_wTwR}
\title{Graph for enumeration of possibilities under Hd for sample-dependent error probabilities, $w_t$ and $w_r$}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_Ha_wTwR
}
\description{
Graph for enumeration of possibilities under Hd for sample-dependent error probabilities, $w_t$ and $w_r$
}
\keyword{datasets}
