% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_Hd}
\alias{graph_Hd}
\title{Graph for enumeration of possibilities under Hd}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_Hd
}
\description{
Graph for enumeration of possibilities under Hd
}
\keyword{datasets}
