% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_Hd_w}
\alias{graph_Hd_w}
\title{Graph for enumeration of possibilities under Hd for one error probability, $w$}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_Hd_w
}
\description{
Graph for enumeration of possibilities under Hd for one error probability, $w$
}
\keyword{datasets}
