% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_Hd_wDwS}
\alias{graph_Hd_wDwS}
\title{Graph for enumeration of possibilities under Hd for sample-dependent error probabilities, $w_D$ and $w_S$}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_Hd_wDwS
}
\description{
Graph for enumeration of possibilities under Hd for sample-dependent error probabilities, $w_D$ and $w_S$
}
\keyword{datasets}
