% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_Hp}
\alias{graph_Hp}
\title{Graph for enumeration of possibilities under Hp}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_Hp
}
\description{
Graph for enumeration of possibilities under Hp
}
\keyword{datasets}
