% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph_single}
\alias{graph_single}
\title{Graph for enumeration of possibilities for single observation}
\format{
An object of class \code{tbl_graph} (inherits from \code{igraph}) of length 10.
}
\usage{
graph_single
}
\description{
Graph for enumeration of possibilities for single observation
}
\keyword{datasets}
