% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_candidates_v1}
\alias{marker_candidates_v1}
\title{Marker candidates (v1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4128 rows and 23 columns.
}
\source{
\url{https://gnomad.broadinstitute.org/}
}
\usage{
marker_candidates_v1
}
\description{
Marker candidates (v1)
}
\keyword{datasets}
