% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Extract posterior samples}
\usage{
posterior_samples(x)
}
\arguments{
\item{x}{Result from \code{\link[=estimate_w_bayesian]{estimate_w_bayesian()}}}
}
\description{
Extract posterior samples
}
\examples{
tab1 <- matrix(c(1000, 10, 2, 12, 100, 8, 1, 7, 200), nrow = 3)
tab1
y <- estimate_w_bayesian(tab1)
q <- posterior_samples(y)
mean(q); #plot(q, type = "l"); hist(q)

}
