% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{posterior_samples_df}
\alias{posterior_samples_df}
\title{Extract posterior samples as data.frame}
\usage{
posterior_samples_df(x)
}
\arguments{
\item{x}{Result from \code{\link[=estimate_w_bayesian]{estimate_w_bayesian()}}}
}
\description{
Extract posterior samples as data.frame
}
\examples{
tab1 <- matrix(c(1000, 10, 2, 12, 100, 8, 1, 7, 200), nrow = 3)
tab1
y <- estimate_w_bayesian(tab1)
q <- posterior_samples_df(y)
head(q)
mean(q$sample); #ggplot(q, aes(iteration, sample, colour = factor(chain))) + geom_line()

}
