% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_data_Ha_wTwR}
\alias{sample_data_Ha_wTwR}
\title{Sample cases under Hd for sample-dependent error probabilities, $w_t$ and $w_r$}
\usage{
sample_data_Ha_wTwR(n, wT, wR, p, ...)
}
\arguments{
\item{n}{number of samples}

\item{wT}{error probability for donor sample}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (length is number of loci)
or vector of length 3 for single locus}

\item{\dots}{Passed on to \code{\link[=add_errors_to_genotypes]{add_errors_to_genotypes()}}}
}
\value{
list of two matrices, each of size n x loci with
genotype in 0/1/2 format resembling the situation in real life.
}
\description{
One latent genotype, ZD, for true donor (D) and
one latent genotype, ZS, for suspect (S).
}
\examples{
sample_data_Ha_wTwR(n = 10, wT = 0.3, wR = 1e-6, p = c(0.25, 0.25, 0.5))
sample_data_Ha_wTwR(n = 10, wT = 0.3, wR = 1e-6, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
  
cases <- sample_data_Ha_wTwR(n = 1000, wT = 1e-1, wR = 1e-8, p = c(0.25, 0.25, 0.5))
tab <- table(xT = cases$xT, xR = cases$xR)
tab

cases <- sample_data_Ha_wTwR(n = 1000, wT = 0, wR = 0, p = c(0.25, 0.25, 0.5))
tab <- table(xT = cases$xT, xR = cases$xR)
tab

}
