% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_data_Hd_w}
\alias{sample_data_Hd_w}
\title{Sample cases under Hd for one error probability, $w$}
\usage{
sample_data_Hd_w(n, w, p, ...)
}
\arguments{
\item{n}{number of samples}

\item{w}{error probability}

\item{p}{list of genotype probabilities (length is number of loci)
or vector of length 3 for single locus}

\item{\dots}{Passed on to \code{\link[=add_errors_to_genotypes]{add_errors_to_genotypes()}}}
}
\value{
list of two matrices, each of size n x loci with
genotype in 0/1/2 format resembling the situation in real life.
}
\description{
One latent genotype, ZD, for true donor (D) and
one latent genotype, ZS, for suspect (S).
}
\examples{
sample_data_Hd_w(n = 10, w = 0.3, p = c(0.25, 0.25, 0.5))
sample_data_Hd_w(n = 10, w = 0.1, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
  
cases <- sample_data_Hd_w(n = 1000, w = 0, p = c(0.25, 0.25, 0.5))
tab <- table(xD = cases$xD, xS = cases$xS)
tab

}
