% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_data_Hp}
\alias{sample_data_Hp}
\title{Sample cases under Hp}
\usage{
sample_data_Hp(n, w, p)
}
\arguments{
\item{n}{number of samples}

\item{w}{error probability}

\item{p}{list of genotype probabilities (length is number of loci)
or vector of length 3 for single locus}
}
\value{
list of two matrices, each of size n x loci with
genotype in 0/1/2 format resembling the situation in real life.
}
\description{
Same latent genotype, Z, with independent errors for true donor (D) and
suspect (S).
}
\examples{
sample_data_Hp(n = 10, w = 0.3, p = c(0.25, 0.25, 0.5))
sample_data_Hp(n = 10, w = 0.1, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
  
cases <- sample_data_Hp(n = 1000, w = 0.3, p = c(0.25, 0.25, 0.5))
tab <- table(cases$X_D, cases$X_S)
tab
estimate_w(tab)

}
