% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_data_Hp_wDwS}
\alias{sample_data_Hp_wDwS}
\title{Sample cases under Hp for sample-dependent error probabilities, $w_D$ and $w_S$}
\usage{
sample_data_Hp_wDwS(n, wD, wS, p, ...)
}
\arguments{
\item{n}{number of samples}

\item{wD}{error probability for donor sample}

\item{wS}{error probability for PoI sample}

\item{p}{list of genotype probabilities (length is number of loci)
or vector of length 3 for single locus}

\item{\dots}{Passed on to \code{\link[=add_errors_to_genotypes]{add_errors_to_genotypes()}}}
}
\value{
list of two matrices, each of size n x loci with
genotype in 0/1/2 format resembling the situation in real life.
}
\description{
Same latent genotype, Z, with independent errors for true donor (D) and
suspect (S).
}
\examples{
sample_data_Hp_wDwS(n = 10, wD = 0.3, wS = 1e-6, p = c(0.25, 0.25, 0.5))
sample_data_Hp_wDwS(n = 10, wD = 0.3, wS = 1e-6, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
  
cases <- sample_data_Hp_wDwS(n = 1000, wD = 0, wS = 0, p = c(0.25, 0.25, 0.5))
table(X_D = cases$X_D, X_S = cases$X_S)
cases <- sample_data_Hp_wDwS(n = 1000, wD = 0.1, wS = 0, p = c(0.25, 0.25, 0.5))
table(X_D = cases$X_D, X_S = cases$X_S)
cases <- sample_data_Hp_wDwS(n = 1000, wD = 0, wS = 0.1, p = c(0.25, 0.25, 0.5))
table(X_D = cases$X_D, X_S = cases$X_S)
  
cases <- sample_data_Hp_wDwS(n = 1000, wD = 1e-1, wS = 1e-8, p = c(0.25, 0.25, 0.5))
tab <- table(X_D = cases$X_D, X_S = cases$X_S)
tab
estimate_w(tab)

cases <- sample_data_Hp_wDwS(n = 1000, wD = 0, wS = 0, p = c(0.1, 0.7, 0.2))
tab <- table(X_D = cases$X_D, X_S = cases$X_S)
diag(tab/sum(tab))

}
