% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sample_profiles_without_error}
\alias{sample_profiles_without_error}
\title{Sample genotype without error}
\usage{
sample_profiles_without_error(n, p)
}
\arguments{
\item{n}{number of samples}

\item{p}{list of genotype probabilities or vector of length 3 for single locus}
}
\value{
list, element for each locus is a matrix with n rows and two columns
}
\description{
Sample genotype without error
}
\examples{
Z <- sample_profiles_without_error(n = 2, p = c(0.25, 0.25, 0.5))
to012(Z)
Z <- sample_profiles_without_error(n = 5, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
Z
to012(Z)

}
