% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to012}
\alias{to012}
\title{Convert to 0/1/2 notation}
\usage{
to012(x)
}
\arguments{
\item{x}{genotypes, e.g. from \code{\link[=sample_profiles_without_error]{sample_profiles_without_error()}}}
}
\description{
Can both
\enumerate{
\item Convert character genotype 0/0, 0/1, 1/0, 1/1 to 0, 1, 2 notation.
\item Convert haplotypes c(0, 0), c(0, 1), c(1, 0), c(1, 1) to 0, 1, 2 notation.
}
}
\details{
Note that input is a list of loci, each element matrix n x 2 where
n is number of individuals.
Output here is n x loci matrix.
}
\examples{
to012(c("0/0", "0/1", "1/0", "1/1"))

Z <- sample_profiles_without_error(n = 10, p = list(
  c(0.25, 0.25, 0.5), c(0.01, 0.01, 0.98)))
Z
to012(Z)

}
