% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wT-posterior-distributions.R
\name{wT_posterior_Ha_naive}
\alias{wT_posterior_Ha_naive}
\title{Sample from posterior distribution of \code{wT} conditional on Ha using a naive MCMC sampler}
\usage{
wT_posterior_Ha_naive(
  xT,
  xR,
  wR,
  p,
  prior_a = 1,
  prior_b = 1,
  chains = 4,
  chains_function = lapply,
  warmup = 1000,
  iterations = 10000,
  ...
)
}
\arguments{
\item{xT}{profile from case (of 0, 1, 2)}

\item{xR}{profile from suspect (of 0, 1, 2)}

\item{wR}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xT}/\code{xR}, or vector of length 3 for reuse)}

\item{prior_a}{First shape parameter for prior beta distribution}

\item{prior_b}{Second shape parameter for prior beta distribution}

\item{chains}{Number of chains to run}

\item{chains_function}{Function to run the chains (fx \code{lapply} or \code{parallel::mclapply})}

\item{warmup}{Number of warm-up iterations}

\item{iterations}{Number of iterations (after warm-up)}

\item{\dots}{Not used}
}
\description{
Sample from posterior distribution of \code{wT} conditional on Ha using a naive MCMC sampler
}
