% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{add_errors_to_genotypes}
\alias{add_errors_to_genotypes}
\title{Add errors to genotypes}
\usage{
add_errors_to_genotypes(Z, w)
}
\arguments{
\item{Z}{genotypes, e.g. created by \code{\link[=sample_profiles_without_error]{sample_profiles_without_error()}}}

\item{w}{error probability}
}
\value{
list, element for each locus is a matrix with n rows and two columns
}
\description{
Add errors to genotypes
}
\examples{
Z <- sample_profiles_without_error(n = 5, p = list(
  c(0.25, 0.25, 0.5), c(0.1, 0.8, 0.1)))
to012(Z)
X <- add_errors_to_genotypes(Z, 0.5)
X
to012(Z) - to012(add_errors_to_genotypes(Z, 0.0))


Z <- sample_profiles_without_error(n = 1000, p = c(0.5, 0.4, 0.1))
X1 <- add_errors_to_genotypes(Z, 0.5)
X2 <- add_errors_to_genotypes(Z, 0.5)
tab <- table(to012(X1), to012(X2))
tab
estimate_w(tab)

}
