% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{calc_LRs_wDwS_integrate_wD_mc}
\alias{calc_LRs_wDwS_integrate_wD_mc}
\title{Calculate LR for a profile for sample-specific error probabilities integrated over the donor prior distribution using Monte Carlo integration}
\usage{
calc_LRs_wDwS_integrate_wD_mc(xD, xS, shape1D, shape2D, wS, p, n_samples = 100)
}
\arguments{
\item{xD}{profile from case (of 0, 1, 2)}

\item{xS}{profile from suspect (of 0, 1, 2)}

\item{shape1D}{\code{wD} has beta prior (with support on 0-0.5) with parameters \code{shape1D} and \code{shape2D}}

\item{shape2D}{see \code{shape1D}}

\item{wS}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xD}/\code{xS}, or vector of length 3 for reuse)}

\item{n_samples}{number of random samples from each prior distribution}
}
\description{
Calculate LR for a profile for sample-specific error probabilities integrated over the donor prior distribution using Monte Carlo integration
}
\examples{
calc_LRs_wDwS(xD = c(0, 0), xS = c(0, 1), wD = 1e-2, wS = 1e-5, p = c(0.25, 0.25, 0.5))

shpD <- get_beta_parameters(mu = 1e-2, sigmasq = 2*1e-3, a = 0, b = 0.5)
calc_LRs_wDwS_integrate_wD_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  wS = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, shpD[1], shpD[2]), from = 0, to = 0.5)
calc_LRs_wDwS_integrate_wD_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = 1, shape2D = 1, 
  wS = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, 1, 1), from = 0, to = 0.5)
  
shpS <- get_beta_parameters(mu = 1e-5, sigmasq = 1e-7, a = 0, b = 0.5)
integrate(function(x) dbeta05(x, shape1 = shpS[1L], shape2 = shpS[2L]), lower = 1e-12, upper = 0.5, rel.tol = 1e-6)
calc_LRs_wDwS_integrate_wDwS_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  shape1S = shpS[1], shape2S = shpS[2], 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, shpS[1], shpS[2]), from = 0, to = 0.5)

}
