% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{calc_LRs_wDwS_integrate_wD_num}
\alias{calc_LRs_wDwS_integrate_wD_num}
\title{Calculate LR for a profile for sample-specific error probabilities integrated over the donor prior distribution using numerical integration}
\usage{
calc_LRs_wDwS_integrate_wD_num(
  xD,
  xS,
  shape1D,
  shape2D,
  wS,
  p,
  n_samples = 100
)
}
\arguments{
\item{xD}{profile from case (of 0, 1, 2)}

\item{xS}{profile from suspect (of 0, 1, 2)}

\item{shape1D}{\code{wD} has beta prior (with support on 0-0.5) with parameters \code{shape1D} and \code{shape2D}}

\item{shape2D}{see \code{shape1D}}

\item{wS}{error probability for PoI sample}

\item{p}{list of genotype probabilities (same length as \code{xD}/\code{xS}, or vector of length 3 for reuse)}
}
\description{
Calculate LR for a profile for sample-specific error probabilities integrated over the donor prior distribution using numerical integration
}
\examples{
calc_LRs_wDwS(xD = c(0, 0), xS = c(0, 1), wD = 1e-2, wS = 1e-5, p = c(0.25, 0.25, 0.5))

shpD <- get_beta_parameters(mu = 1e-2, sigmasq = 2*1e-3, a = 0, b = 0.5)
calc_LRs_wDwS_integrate_wD_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  wS = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
calc_LRs_wDwS_integrate_wD_num(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = shpD[1], shape2D = shpD[2], 
  wS = 1e-5, 
  p = c(0.25, 0.25, 0.5))
# curve(dbeta05(x, shpD[1], shpD[2]), from = 0, to = 0.5)

calc_LRs_wDwS_integrate_wD_mc(
  xD = c(0, 0), 
  xS = c(0, 1), 
  shape1D = 1, shape2D = 1, 
  wS = 1e-5, 
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
calc_LRs_wDwS_integrate_wD_num(
  xD = c(0, 0),
  xS = c(0, 1),
  shape1D = 1, shape2D = 1,
  wS = 1e-5,
  p = c(0.25, 0.25, 0.5),
  n_samples = 100)
# curve(dbeta05(x, 1, 1), from = 0, to = 0.5)
  
}
