% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_candidates_v1}
\alias{marker_candidates_v1}
\title{Marker candidates (v1)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4128 rows and 23 columns.
}
\source{
\url{https://gnomad.broadinstitute.org/}
}
\usage{
marker_candidates_v1
}
\description{
Data from The Genome Aggregation Database (gnomAD) for
binary SNP loci with \code{rs} names, i.e.,
where the \code{rs} column starts with \code{rs},
there is one character in the \code{ref} column and one character in the \code{alt} column.
}
\details{
The loci are filtered by MAF values for AFR, EAS, and NFE populations.
The MAF values must be within 0.1, 0.15, 0.2 from 0.5.
}
\keyword{datasets}
